Fix failing benchmark_count_sum.py
Apparently the benchmark job that I thought spuriously timed out actually failed with an exception in !173

No changelog entry b/c this is fixing something that hasn't been released yet

Link to Issue: tumult#2491, tumult#2565

### For the Assignee:
- [X] Update the UNRELEASED sections of the changelogs for each project modified, 
        or explain in the MR description why no updates were necessary.

### For the Reviewer:
- [ ] Confirm that an Issue is linked, if applicable.
- [ ] Confirm that the assignee updated relevant changelogs or provided appropriate
    justification as to why no updates were necessary.

This template lives [here](https://gitlab.com/tumult-labs/core/-/blob/dev/.gitlab/merge_request_templates/default_template.md)

+++ benchmark/benchmark_count_sum.py
@@ -12,7 +12,7 @@ import pandas as pd
 from pyspark.sql import SparkSession
 from pyspark.sql import functions as sf
 from pyspark.sql.dataframe import DataFrame
-from pyspark.sql.types import IntegerType, StructField, StructType
+from pyspark.sql.types import LongType, StructField, StructType
 
 from benchmarking_utils import Timer, write_as_html
 from tmlt.core.domains.spark_domains import (
@@ -240,7 +240,7 @@ def main():
         )
         schema = StructType(
             [
-                StructField("Col_{}".format(i), IntegerType(), True)
+                StructField("Col_{}".format(i), LongType(), True)
                 for i in range(num_cols)
             ]
         )
@@ -292,7 +292,7 @@ def main():
         )
         schema = StructType(
             [
-                StructField("Col_{}".format(i), IntegerType(), True)
+                StructField("Col_{}".format(i), LongType(), True)
                 for i in range(num_cols)
             ]
         )
